src/requirements.txt
fastapi==0.79.0
python-kasa==0.5.0
prometheus-client==0.14.1
apscheduler==3.9.1
uvicorn==0.18.2